requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

# Vector Database and Embeddings
chromadb==0.4.22
hnswlib==0.8.0

# Document Processing
PyPDF2==3.0.1
python-magic==0.4.27

# Audio/Video Processing
openai-whisper==20231117
yt-dlp==2023.12.30
ffmpeg-python==0.2.0

# HTTP and Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
numpy==1.24.3
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
uuid==1.30

# Type Hints
typing-extensions==4.9.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code Analysis (uses built-in Python 're' module)
# No additional dependencies required for code review functionality

# Optional: Performance
numexpr==2.8.8